+++ docs/release-notes.md
@@ -1,6 +1,18 @@
 # msaDocModels Release Notes
 ## Possible future features:
 
+# 0.0.98
+
+- fix algorithm to DocClassifier Input 
+
+# 0.0.97
+
+- add algorithm to DocClassifier Input 
+- change dto model for DocClassifier
+
+# 0.0.96
+
+- add include_fields field to CreatePDFInputModel
 
 # 0.0.95